updated installation file

+++ install.sh
@@ -15,28 +15,7 @@ else
 	cd ../
 	if  test -f "./src/conda/bin/conda"; then
 	echo "Miniconda3 successfully installed"
-	fi
-fi
-
-
-
-if  test -f "./src/conda2/bin/conda"; then
-	echo "Miniconda2 already installed"
-else
-
-
-	cd src
-	bash Miniconda2-latest-Linux-x86_64.sh -b -p ./conda2
-	cd ../
-	if  test -f "./src/conda2/bin/conda"; then
-	echo "Miniconda2 successfully installed"
-	fi
-fi
-
-#Check if a c compiler is installed. If not:
-# sudo apt-get install build-essential
-
-./src/conda/bin/conda install -c anaconda -y pillow
+	./src/conda/bin/conda install -c anaconda -y pillow
 ./src/conda/bin/conda install -c anaconda -y numpy
 ./src/conda/bin/conda install -c conda-forge -y matplotlib 
 ./src/conda/bin/conda config --add channels bioconda
@@ -57,7 +36,6 @@ fi
 ./src/conda/bin/conda install -c bioconda -y  spades=3.12
 ./src/conda/bin/conda install -c bioconda -y  picard=2.21
 ./src/conda/bin/conda install -c bioconda -y  lastz=1.0.4
-./src/conda2/bin/conda install -c bioconda -y  ragout=2.2
 ./src/conda/bin/conda install -c bioconda -y  perl-perl4-corelibs
 ./src/conda/bin/conda install -c bioconda -y  blast=2.9.0
 ./src/conda/bin/conda install -c bioconda -y  lofreq=2.1.4
@@ -66,11 +44,35 @@ fi
 ./src/conda/bin/conda install -c bioconda -y  bedtools=2.29.2
 ./src/conda/bin/conda install -c bioconda -y  fastx_toolkit=0.0.14
 ./src/conda/bin/conda install -c bioconda -y  blat=36
-./src/conda2/bin/conda install -c bioconda -y  mummer
 ./src/conda/bin/conda install -c bioconda -y  bcftools=1.8
 ./src/conda/bin/conda install -c bioconda -y  exonerate=2.4
 ./src/conda/bin/conda install -c anaconda -y pyqt=5.9.2
 ./src/conda/bin/conda install -c bioconda -y varscan=2.4.4
+	fi
+fi
+
+
+
+if  test -f "./src/conda2/bin/conda"; then
+	echo "Miniconda2 already installed"
+else
+
+
+	cd src
+	bash Miniconda2-latest-Linux-x86_64.sh -b -p ./conda2
+	cd ../
+	if  test -f "./src/conda2/bin/conda"; then
+	echo "Miniconda2 successfully installed"
+	./src/conda2/bin/conda install -c bioconda -y  ragout=2.2
+	./src/conda2/bin/conda install -c bioconda -y  mummer
+
+	fi
+fi
+
+#Check if a c compiler is installed. If not:
+# sudo apt-get install build-essential
+
+
